requirements.txt
flask
gunicorn
torch
torchvision
pillow
gdown